requirements.txt
python-whois
python-dotenv
requests
python-telegram-bot
tldextract
